requirements.txt
paho-mqtt==1.3.1
typing==3.6.2
ruamel.yaml==0.15.34
python-slugify==1.2.4
wakeonlan==1.1.6